Accept a --show-config option
This will be used to show the user what the current configuration is.  One
can specify a config using the relevant file, and one can override the
configuration using command line options.  The --show-config option should
then be able to show the effects of setting an option on the command line on
a preset config option.